Improve SSTable handling and persistence.   - Add SSTable::create() and SSTable::open() factory methods (return std::expected instead of throwing)   - Add lsm.meta file to track SSTable ordering   - Fix MemTable::flush_to_disk to use write() instead of << (handles binary data correctly)   - Fix LsmTree::get() expected/optional handling bug   - Update tests for new SSTable API

+++ src/Constants.h
@@ -1,7 +1,8 @@
 #pragma once
+#include <cstddef>
 namespace lsm_storage_engine {
 namespace lsm_constants {
 // TODO: Add to configuration.
-constexpr unsigned long kMemTableFlushThreshold = 1UZ << 16;
+constexpr size_t kMemTableFlushThreshold = 1UZ << 16;
 } // namespace lsm_constants
 }; // namespace lsm_storage_engine
